requirements.txt
torch>=2.0
torchaudio>=2.0
torchvision>=0.15

autrainer[all]

librosa>=0.10
audiomentations>=0.31.0
torch-audiomentations>=0.11.0
soundfile
pydub

pandas
numpy
scipy

matplotlib
seaborn
tensorboard

mlflow

tqdm
omegaconf
hydra-core>=1.3

ffmpeg-python

python-dotenv
wandb>=0.16
